Split pytato.distributed up into submodules (no functional changes)
