+++ pytato/distributed/partition.py
@@ -1,3 +1,12 @@
+"""
+.. currentmodule:: pytato
+
+.. autoclass:: DistributedGraphPart
+.. autoclass:: DistributedGraphPartition
+
+.. autofunction:: find_distributed_partition
+"""
+
 from __future__ import annotations
 
 __copyright__ = """
@@ -24,308 +33,28 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 """
 
-from typing import (Any, Dict, Hashable, Tuple, Optional, Set,  # noqa: F401
-                    List, FrozenSet, Callable, cast, Mapping, Iterable, Sequence,
-                    ClassVar, TYPE_CHECKING
-                    )  # Mapping required by sphinx
-from immutables import Map
+from typing import (
+        Tuple, Any, Mapping, FrozenSet, Set, Dict, cast, Iterable, Callable, List)
+from functools import cached_property
 
 import attrs
+from immutables import Map
 
+from pymbolic.mapper.optimize import optimize_mapper
 from pytools import UniqueNameGenerator
-from pytools.tag import Taggable, UniqueTag, Tag
-from pytato.array import (Array, _SuppliedShapeAndDtypeMixin,
-                          DictOfNamedArrays, ShapeType, Placeholder,
-                          make_placeholder, _get_default_axes, AxesT,
+from pytools.tag import UniqueTag
+
+from pytato.scalar_expr import SCALAR_CLASSES
+from pytato.array import (Array,
+                          DictOfNamedArrays, Placeholder, make_placeholder,
                           NamedArray)
 from pytato.transform import (ArrayOrNames, CopyMapper, Mapper,
                               CachedWalkMapper, CopyMapperWithExtraArgs,
                               CombineMapper)
-
 from pytato.partition import GraphPart, GraphPartition, PartId, GraphPartitioner
-from pytato.target import BoundProgram
+from pytato.distributed.nodes import (
+        DistributedRecv, DistributedSend, DistributedSendRefHolder)
 from pytato.analysis import DirectPredecessorsGetter
-from functools import cached_property
-from pytato.scalar_expr import SCALAR_CLASSES, INT_CLASSES
-from pymbolic.mapper.optimize import optimize_mapper
-
-import numpy as np
-
-import logging
-logger = logging.getLogger(__name__)
-
-
-if TYPE_CHECKING:
-    import mpi4py.MPI
-
-
-__doc__ = r"""
-Distributed-memory evaluation of expression graphs is accomplished
-by :ref:`partitioning <partitioning>` the graph to reveal communication-free
-pieces of the computation. Communication (i.e. sending/receiving data) is then
-accomplished at the boundaries of the parts of the resulting graph partitioning.
-
-Recall the requirement for partitioning that, "no part may depend on its own
-outputs as inputs". That sounds obvious, but in the distributed-memory case,
-this is harder to decide than it looks, since we do not have full knowledge of
-the computation graph.  Edges go off to other nodes and then come back.
-
-As a first step towards making this tractable, we currently strengthen the
-requirement to create partition boundaries on every edge that goes between
-nodes that are/are not a dependency of a receive or that feed/do not feed a send.
-
-.. currentmodule:: pytato
-.. autoclass:: DistributedSend
-.. autoclass:: DistributedSendRefHolder
-.. autoclass:: DistributedRecv
-
-.. autofunction:: make_distributed_send
-.. autofunction:: staple_distributed_send
-.. autofunction:: make_distributed_recv
-
-.. currentmodule:: pytato.distributed
-
-.. autoclass:: DistributedGraphPart
-.. autoclass:: DistributedGraphPartition
-.. autofunction:: verify_distributed_partition
-
-.. currentmodule:: pytato
-
-.. autofunction:: find_distributed_partition
-.. autofunction:: number_distributed_tags
-.. autofunction:: execute_distributed_partition
-
-Internal stuff that is only here because the documentation tool wants it
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-.. class:: Tag
-
-    See :class:`pytools.tag.Tag`.
-
-.. class:: CommTagType
-
-    A type representing a communication tag.
-
-.. class:: ShapeType
-
-    A type representing a shape.
-
-.. class:: AxesT
-
-    A :class:`tuple` of :class:`Axis` objects.
-"""
-
-
-# {{{ distributed node types
-
-CommTagType = Hashable
-
-
-class DistributedSend(Taggable):
-    """Class representing a distributed send operation.
-
-    .. attribute:: data
-
-        The :class:`~pytato.Array` to be sent.
-
-    .. attribute:: dest_rank
-
-        An :class:`int`. The rank to which :attr:`data` is to be sent.
-
-    .. attribute:: comm_tag
-
-        A hashable, picklable object to serve as a 'tag' for the communication.
-        Only a :class:`DistributedRecv` with the same tag will be able to
-        receive the data being sent here.
-    """
-
-    def __init__(self, data: Array, dest_rank: int, comm_tag: CommTagType,
-                 tags: FrozenSet[Tag] = frozenset()) -> None:
-        super().__init__(tags=tags)
-        self.data = data
-        self.dest_rank = dest_rank
-        self.comm_tag = comm_tag
-
-    def __hash__(self) -> int:
-        return (
-                hash(self.__class__)
-                ^ hash(self.data)
-                ^ hash(self.dest_rank)
-                ^ hash(self.comm_tag)
-                ^ hash(self.tags)
-                )
-
-    def __eq__(self, other: Any) -> bool:
-        return (
-                self.__class__ is other.__class__
-                and self.data == other.data
-                and self.dest_rank == other.dest_rank
-                and self.comm_tag == other.comm_tag
-                and self.tags == other.tags)
-
-    def _with_new_tags(self, tags: FrozenSet[Tag]) -> DistributedSend:
-        return DistributedSend(
-                data=self.data,
-                dest_rank=self.dest_rank,
-                comm_tag=self.comm_tag,
-                tags=tags)
-
-    def copy(self, **kwargs: Any) -> DistributedSend:
-        data: Optional[Array] = kwargs.get("data")
-        dest_rank: Optional[int] = kwargs.get("dest_rank")
-        comm_tag: Optional[CommTagType] = kwargs.get("comm_tag")
-        tags = cast(FrozenSet[Tag], kwargs.get("tags"))
-        return type(self)(
-                data=data if data is not None else self.data,
-                dest_rank=dest_rank if dest_rank is not None else self.dest_rank,
-                comm_tag=comm_tag if comm_tag is not None else self.comm_tag,
-                tags=tags if tags is not None else self.tags)
-
-    def __repr__(self) -> str:
-        # self.data takes a lot of space, shorten it
-        return (f"DistributedSend(data={self.data.__class__} "
-                f"at {hex(id(self.data))}, "
-                f"dest_rank={self.dest_rank}, "
-                f"tags={self.tags}, comm_tag={self.comm_tag})")
-
-
-@attrs.define(frozen=True, eq=False, repr=False, init=False)
-class DistributedSendRefHolder(Array):
-    """A node acting as an identity on :attr:`passthrough_data` while also holding
-    a reference to a :class:`DistributedSend` in :attr:`send`. Since
-    :mod:`pytato` represents data flow, and since no data flows 'out'
-    of a :class:`DistributedSend`, no node in all of :mod:`pytato` has
-    a good reason to hold a reference to a send node, since there is
-    no useful result of a send (at least of an :class:`~pytato.Array` type).
-
-    This is where this node type comes in. Its value is the same as that of
-    :attr:`passthrough_data`, *and* it holds a reference to the send node.
-
-    .. note::
-
-        This all seems a wee bit inelegant, but nobody who has written
-        or reviewed this code so far had a better idea. If you do, please speak up!
-
-    .. attribute:: send
-
-        The :class:`DistributedSend` to which a reference is to be held.
-
-    .. attribute:: passthrough_data
-
-        A :class:`~pytato.Array`. The value of this node.
-
-    .. note::
-
-        It is the user's responsibility to ensure matching sends and receives
-        are part of the computation graph on all ranks. If this rule is not heeded,
-        undefined behavior (in particular deadlock) may result.
-        Notably, by the nature of the data flow graph built by :mod:`pytato`,
-        unused results do not appear in the graph. It is thus possible for a
-        :class:`DistributedSendRefHolder` to be constructed and yet to not
-        become part of the graph constructed by the user.
-    """
-    send: DistributedSend
-    passthrough_data: Array
-
-    _mapper_method: ClassVar[str] = "map_distributed_send_ref_holder"
-    _fields: ClassVar[Tuple[str, ...]] = Array._fields + ("passthrough_data", "send")
-
-    def __init__(self, send: DistributedSend, passthrough_data: Array,
-                 tags: FrozenSet[Tag] = frozenset()) -> None:
-        super().__init__(axes=passthrough_data.axes, tags=tags)
-        object.__setattr__(self, "send", send)
-        object.__setattr__(self, "passthrough_data", passthrough_data)
-
-    @property
-    def shape(self) -> ShapeType:
-        return self.passthrough_data.shape
-
-    @property
-    def dtype(self) -> np.dtype[Any]:
-        return self.passthrough_data.dtype
-
-    def copy(self, **kwargs: Any) -> DistributedSendRefHolder:
-        # override 'Array.copy' because
-        # 'DistributedSendRefHolder.axes' is a read-only field.
-        send = kwargs.pop("send", self.send)
-        passthrough_data = kwargs.pop("passthrough_data", self.passthrough_data)
-        tags = kwargs.pop("tags", self.tags)
-
-        if kwargs:
-            raise ValueError("Cannot assign"
-                             f" DistributedSendRefHolder.'{set(kwargs)}'")
-        return DistributedSendRefHolder(send,
-                                        passthrough_data,
-                                        tags)
-
-
-@attrs.define(frozen=True, eq=False)
-class DistributedRecv(_SuppliedShapeAndDtypeMixin, Array):
-    """Class representing a distributed receive operation.
-
-    .. attribute:: src_rank
-
-        An :class:`int`. The rank from which an array is to be received.
-
-    .. attribute:: comm_tag
-
-        A hashable, picklable object to serve as a 'tag' for the communication.
-        Only a :class:`DistributedSend` with the same tag will be able to
-        send the data being received here.
-
-    .. attribute:: shape
-    .. attribute:: dtype
-
-    .. note::
-
-        It is the user's responsibility to ensure matching sends and receives
-        are part of the computation graph on all ranks. If this rule is not heeded,
-        undefined behavior (in particular deadlock) may result.
-        Notably, by the nature of the data flow graph built by :mod:`pytato`,
-        unused results do not appear in the graph. It is thus possible for a
-        :class:`DistributedRecv` to be constructed and yet to not become part
-        of the graph constructed by the user.
-    """
-    src_rank: int
-    comm_tag: CommTagType
-    _shape: ShapeType
-    _dtype: Any  # FIXME: sphinx does not like `_dtype: _dtype_any`
-
-    _fields: ClassVar[Tuple[str, ...]] = Array._fields + ("shape", "dtype",
-                                                          "src_rank", "comm_tag")
-    _mapper_method: ClassVar[str] = "map_distributed_recv"
-
-
-def make_distributed_send(sent_data: Array, dest_rank: int, comm_tag: CommTagType,
-                          send_tags: FrozenSet[Tag] = frozenset()) -> \
-         DistributedSend:
-    """Make a :class:`DistributedSend` object."""
-    return DistributedSend(sent_data, dest_rank, comm_tag, send_tags)
-
-
-def staple_distributed_send(sent_data: Array, dest_rank: int, comm_tag: CommTagType,
-                          stapled_to: Array, *,
-                          send_tags: FrozenSet[Tag] = frozenset(),
-                          ref_holder_tags: FrozenSet[Tag] = frozenset()) -> \
-         DistributedSendRefHolder:
-    """Make a :class:`DistributedSend` object wrapped in a
-    :class:`DistributedSendRefHolder` object."""
-    return DistributedSendRefHolder(
-            DistributedSend(sent_data, dest_rank, comm_tag, send_tags),
-            stapled_to, tags=ref_holder_tags)
-
-
-def make_distributed_recv(src_rank: int, comm_tag: CommTagType,
-                          shape: ShapeType, dtype: Any,
-                          axes: Optional[AxesT] = None,
-                          tags: FrozenSet[Tag] = frozenset()
-                          ) -> DistributedRecv:
-    """Make a :class:`DistributedRecv` object."""
-    if axes is None:
-        axes = _get_default_axes(len(shape))
-    dtype = np.dtype(dtype)
-    return DistributedRecv(src_rank, comm_tag, shape, dtype, tags=tags, axes=axes)
-
-# }}}
 
 
 # {{{ distributed graph partition
@@ -502,7 +231,6 @@ class _DistributedGraphPartitioner(GraphPartitioner):
                 self, expr: DistributedSendRefHolder, *args: Any) -> Any:
         send_part_id = self.get_part_id(expr.send.data)
 
-        from pytato.distributed import DistributedSend
         self.pid_to_dist_sends.setdefault(send_part_id, []).append(
                 DistributedSend(
                     data=self.rec(expr.send.data),
@@ -1035,472 +763,6 @@ def find_distributed_partition(outputs: DictOfNamedArrays
         return send.copy(data=cmac(send.data))
 
     return _map_distributed_graph_partition_nodes(map_array, map_send, gp)
-
-# }}}
-
-
-# {{{ verify_distributed_partition
-
-@attrs.define(frozen=True)
-class _SummarizedDistributedSend:
-    src_rank: int
-    dest_rank: int
-    comm_tag: CommTagType
-
-    shape: ShapeType
-    dtype: np.dtype[Any]
-
-
-@attrs.define(frozen=True)
-class _DistributedPartId:
-    rank: int
-    part_id: PartId
-
-
-@attrs.define(frozen=True)
-class _DistributedName:
-    rank: int
-    name: str
-
-
-@attrs.define(frozen=True)
-class _SummarizedDistributedGraphPart:
-    pid: _DistributedPartId
-    needed_pids: FrozenSet[_DistributedPartId]
-    user_input_names: FrozenSet[_DistributedName]
-    partition_input_names: FrozenSet[_DistributedName]
-    output_names: FrozenSet[_DistributedName]
-    input_name_to_recv_node: Dict[_DistributedName, DistributedRecv]
-    output_name_to_send_node: Dict[_DistributedName, _SummarizedDistributedSend]
-
-    @property
-    def rank(self) -> int:
-        return self.pid.rank
-
-
-@attrs.define(frozen=True)
-class _CommIdentifier:
-    src_rank: int
-    dest_rank: int
-    comm_tag: CommTagType
-
-
-class DistributedPartitionVerificationError(ValueError):
-    pass
-
-
-class DuplicateSendError(DistributedPartitionVerificationError):
-    pass
-
-
-class DuplicateRecvError(DistributedPartitionVerificationError):
-    pass
-
-
-class MissingSendError(DistributedPartitionVerificationError):
-    pass
-
-
-class MissingRecvError(DistributedPartitionVerificationError):
-    pass
-
-
-def verify_distributed_partition(mpi_communicator: mpi4py.MPI.Comm,
-        partition: DistributedGraphPartition) -> None:
-    """
-    Verify that
-
-    - a feasible execution order exists among graph parts across the global,
-      partitioned, distributed data flow graph, consisting of all values of
-      *partition* across all ranks.
-    - sends and receives for a given triple of
-      **(source rank, destination rank, tag)** are unique.
-    - there is a one-to-one mapping between instances of :class:`DistributedRecv`
-      and :class:`DistributedSend`
-
-    .. warning::
-
-        This is an MPI-collective operation.
-    """
-    my_rank = mpi_communicator.rank
-    root_rank = 0
-
-    # Convert local partition to _SummarizedDistributedGraphPart
-    summarized_parts: \
-            Dict[_DistributedPartId, _SummarizedDistributedGraphPart] = {}
-
-    for pid, part in partition.parts.items():
-        assert pid == part.pid
-
-        dpid = _DistributedPartId(my_rank, part.pid)
-        summarized_parts[dpid] = _SummarizedDistributedGraphPart(
-            pid=dpid,
-            needed_pids=frozenset([_DistributedPartId(my_rank, pid)
-                            for pid in part.needed_pids]),
-            user_input_names=frozenset([_DistributedName(my_rank, name)
-                            for name in part.user_input_names]),
-            partition_input_names=frozenset([_DistributedName(my_rank, name)
-                            for name in part.partition_input_names]),
-            output_names=frozenset([_DistributedName(my_rank, name)
-                            for name in part.output_names]),
-            input_name_to_recv_node={_DistributedName(my_rank, name): recv
-                for name, recv in part.input_name_to_recv_node.items()},
-            output_name_to_send_node={
-                _DistributedName(my_rank, name):
-                _SummarizedDistributedSend(
-                            src_rank=my_rank,
-                            dest_rank=send.dest_rank,
-                            comm_tag=send.comm_tag,
-                            shape=send.data.shape,
-                            dtype=send.data.dtype)
-                for name, send in part.output_name_to_send_node.items()})
-
-    # Gather the _SummarizedDistributedGraphPart's to rank 0
-    all_summarized_parts_gathered: Optional[
-            Sequence[Dict[_DistributedPartId, _SummarizedDistributedGraphPart]]] = \
-            mpi_communicator.gather(summarized_parts, root=root_rank)
-
-    if mpi_communicator.rank == root_rank:
-        assert all_summarized_parts_gathered
-
-        all_summarized_parts = {
-                dpid: sumpart
-                for rank_parts in all_summarized_parts_gathered
-                for dpid, sumpart in rank_parts.items()}
-
-        # Every node in the graph is a _SummarizedDistributedGraphPart
-        pid_to_needed_pids: Dict[_DistributedPartId, Set[_DistributedPartId]] = {}
-
-        def add_needed_pid(pid: _DistributedPartId,
-                           needed_pid: _DistributedPartId) -> None:
-            pid_to_needed_pids.setdefault(pid, set()).add(needed_pid)
-
-        all_recvs: Set[_CommIdentifier] = set()
-
-        # {{{ gather information on who produces output
-
-        output_to_defining_pid: Dict[_DistributedName, _DistributedPartId] = {}
-        for sumpart in all_summarized_parts.values():
-            for out_name in sumpart.output_names:
-                assert out_name not in output_to_defining_pid
-                output_to_defining_pid[out_name] = sumpart.pid
-
-        # }}}
-
-        # {{{ gather information on senders
-
-        comm_id_to_sending_pid: Dict[_CommIdentifier, _DistributedPartId] = {}
-        for sumpart in all_summarized_parts.values():
-            for sumsend in sumpart.output_name_to_send_node.values():
-                comm_id = _CommIdentifier(
-                        src_rank=sumsend.src_rank,
-                        dest_rank=sumsend.dest_rank,
-                        comm_tag=sumsend.comm_tag)
-
-                if comm_id in comm_id_to_sending_pid:
-                    raise DuplicateSendError(
-                            f"duplicate send for comm id: '{comm_id}'")
-                comm_id_to_sending_pid[comm_id] = sumpart.pid
-
-        # }}}
-
-        # {{{ add edges to the DAG of distributed graph parts
-
-        for sumpart in all_summarized_parts.values():
-            pid_to_needed_pids[sumpart.pid] = set(sumpart.needed_pids)
-
-            # Loop through all receives, assert that combination of
-            # (src_rank, dest_rank, tag) is unique.
-            for dname, dist_recv in sumpart.input_name_to_recv_node.items():
-                comm_id = _CommIdentifier(
-                        src_rank=dist_recv.src_rank,
-                        dest_rank=dname.rank,
-                        comm_tag=dist_recv.comm_tag)
-
-                if comm_id in all_recvs:
-                    raise DuplicateRecvError(f"Duplicate recv: '{comm_id}'")
-
-                all_recvs.add(comm_id)
-
-                # Add edges between sends and receives (cross-rank)
-                try:
-                    sending_pid = comm_id_to_sending_pid[comm_id]
-                except KeyError:
-                    raise MissingSendError(
-                        f"no matching send for recv on '{comm_id}'")
-
-                add_needed_pid(sumpart.pid, sending_pid)
-
-            # Add edges between output_names and partition_input_names (intra-rank)
-            for input_name in sumpart.partition_input_names:
-                # Input names from recv nodes have no corresponding output_name
-                if input_name in sumpart.input_name_to_recv_node.keys():
-                    continue
-                defining_pid = output_to_defining_pid[input_name]
-                assert defining_pid.rank == sumpart.pid.rank
-                add_needed_pid(sumpart.pid, defining_pid)
-
-        # }}}
-
-        # Loop through all sends again, making sure there exists a matching recv
-        for s in comm_id_to_sending_pid:
-            if s not in all_recvs:
-                raise MissingRecvError(f"no matching recv for send: {s=}")
-
-        # Do a topological sort to check for any cycles
-
-        from pytools.graph import compute_topological_order, CycleError
-        from pytato.partition import PartitionInducedCycleError
-        try:
-            compute_topological_order(pid_to_needed_pids)
-        except CycleError:
-            raise PartitionInducedCycleError
-
-        logger.info("verify_distributed_partition completed successfully.")
-
-# }}}
-
-
-# {{{ construct tag numbering
-
-def number_distributed_tags(
-        mpi_communicator: mpi4py.MPI.Comm,
-        partition: DistributedGraphPartition,
-        base_tag: int) -> Tuple[DistributedGraphPartition, int]:
-    """Return a new :class:`~pytato.distributed.DistributedGraphPartition`
-    in which symbolic tags are replaced by unique integer tags, created by
-    counting upward from *base_tag*.
-
-    :returns: a tuple ``(partition, next_tag)``, where *partition* is the new
-        :class:`~pytato.distributed.DistributedGraphPartition` with numerical
-        tags, and *next_tag* is the lowest integer tag above *base_tag* that
-        was not used.
-
-    .. note::
-
-        This is a potentially heavyweight MPI-collective operation on
-        *mpi_communicator*.
-    """
-    tags = frozenset({
-            recv.comm_tag
-            for part in partition.parts.values()
-            for recv in part.input_name_to_recv_node.values()
-            } | {
-            send.comm_tag
-            for part in partition.parts.values()
-            for send in part.output_name_to_send_node.values()})
-
-    from mpi4py import MPI
-
-    def set_union(
-            set_a: FrozenSet[Any], set_b: FrozenSet[Any],
-            mpi_data_type: MPI.Datatype) -> FrozenSet[str]:
-        assert mpi_data_type is None
-        assert isinstance(set_a, frozenset)
-        assert isinstance(set_b, frozenset)
-
-        return set_a | set_b
-
-    root_rank = 0
-
-    set_union_mpi_op = MPI.Op.Create(
-            # type ignore reason: mpi4py misdeclares op functions as returning
-            # None.
-            set_union,  # type: ignore[arg-type]
-            commute=True)
-    try:
-        all_tags = mpi_communicator.reduce(
-                tags, set_union_mpi_op, root=root_rank)
-    finally:
-        set_union_mpi_op.Free()
-
-    if mpi_communicator.rank == root_rank:
-        sym_tag_to_int_tag = {}
-        next_tag = base_tag
-        assert isinstance(all_tags, frozenset)
-
-        for sym_tag in all_tags:
-            sym_tag_to_int_tag[sym_tag] = next_tag
-            next_tag += 1
-
-        mpi_communicator.bcast((sym_tag_to_int_tag, next_tag), root=root_rank)
-    else:
-        sym_tag_to_int_tag, next_tag = mpi_communicator.bcast(None, root=root_rank)
-
-    from attrs import evolve as replace
-    return DistributedGraphPartition(
-            parts={
-                pid: replace(part,
-                    input_name_to_recv_node={
-                        name: recv.copy(comm_tag=sym_tag_to_int_tag[recv.comm_tag])
-                        for name, recv in part.input_name_to_recv_node.items()},
-                    output_name_to_send_node={
-                        name: send.copy(comm_tag=sym_tag_to_int_tag[send.comm_tag])
-                        for name, send in part.output_name_to_send_node.items()},
-                    )
-                for pid, part in partition.parts.items()
-                },
-            var_name_to_result=partition.var_name_to_result,
-            toposorted_part_ids=partition.toposorted_part_ids), next_tag
-
-# }}}
-
-
-# {{{ distributed execute
-
-def _post_receive(mpi_communicator: mpi4py.MPI.Comm,
-                 recv: DistributedRecv) -> Tuple[Any, np.ndarray[Any, Any]]:
-    if not all(isinstance(dim, INT_CLASSES) for dim in recv.shape):
-        raise NotImplementedError("Parametric shapes not supported yet.")
-
-    assert isinstance(recv.comm_tag, int)
-    # mypy is right here, size params in 'recv.shape' must be evaluated
-    buf = np.empty(recv.shape, dtype=recv.dtype)  # type: ignore[arg-type]
-
-    return mpi_communicator.Irecv(
-            buf=buf, source=recv.src_rank, tag=recv.comm_tag), buf
-
-
-def _mpi_send(mpi_communicator: Any, send_node: DistributedSend,
-             data: np.ndarray[Any, Any]) -> Any:
-    # Must use-non-blocking send, as blocking send may wait for a corresponding
-    # receive to be posted (but if sending to self, this may only occur later).
-    return mpi_communicator.Isend(
-            data, dest=send_node.dest_rank, tag=send_node.comm_tag)
-
-
-def execute_distributed_partition(
-        partition: DistributedGraphPartition, prg_per_partition:
-        Dict[Hashable, BoundProgram],
-        queue: Any, mpi_communicator: Any,
-        *,
-        allocator: Optional[Any] = None,
-        input_args: Optional[Dict[str, Any]] = None) -> Dict[str, Any]:
-
-    if input_args is None:
-        input_args = {}
-
-    from mpi4py import MPI
-
-    if len(partition.parts) != 1:
-        recv_names_tup, recv_requests_tup, recv_buffers_tup = zip(*[
-            (name,) + _post_receive(mpi_communicator, recv)
-            for part in partition.parts.values()
-            for name, recv in part.input_name_to_recv_node.items()])
-        recv_names = list(recv_names_tup)
-        recv_requests = list(recv_requests_tup)
-        recv_buffers = list(recv_buffers_tup)
-        del recv_names_tup
-        del recv_requests_tup
-        del recv_buffers_tup
-    else:
-        # Only a single partition, no recv requests exist
-        recv_names = []
-        recv_requests = []
-        recv_buffers = []
-
-    context: Dict[str, Any] = input_args.copy()
-
-    pids_to_execute = set(partition.parts)
-    pids_executed = set()
-    recv_names_completed = set()
-    send_requests = []
-
-    # {{{ Input name refcount
-
-    # Keep a count on how often each input name is used
-    # in order to be able to free them.
-
-    from pytools import memoize_on_first_arg
-
-    @memoize_on_first_arg
-    def _get_partition_input_name_refcount(partition: DistributedGraphPartition) \
-            -> Dict[str, int]:
-        partition_input_names_refcount: Dict[str, int] = {}
-        for pid in set(partition.parts):
-            for name in partition.parts[pid].all_input_names():
-                if name in partition_input_names_refcount:
-                    partition_input_names_refcount[name] += 1
-                else:
-                    partition_input_names_refcount[name] = 1
-
-        return partition_input_names_refcount
-
-    partition_input_names_refcount = \
-        _get_partition_input_name_refcount(partition).copy()
-
-    # }}}
-
-    def exec_ready_part(part: DistributedGraphPart) -> None:
-        inputs = {k: context[k] for k in part.all_input_names()}
-
-        _evt, result_dict = prg_per_partition[part.pid](queue,
-                                                        allocator=allocator,
-                                                        **inputs)
-
-        context.update(result_dict)
-
-        for name, send_node in part.output_name_to_send_node.items():
-            # FIXME: pytato shouldn't depend on pyopencl
-            if isinstance(context[name], np.ndarray):
-                data = context[name]
-            else:
-                data = context[name].get(queue)
-            send_requests.append(_mpi_send(mpi_communicator, send_node, data))
-
-        pids_executed.add(part.pid)
-        pids_to_execute.remove(part.pid)
-
-    def wait_for_some_recvs() -> None:
-        complete_recv_indices = MPI.Request.Waitsome(recv_requests)
-
-        # Waitsome is allowed to return None
-        if not complete_recv_indices:
-            complete_recv_indices = []
-
-        # reverse to preserve indices
-        for idx in sorted(complete_recv_indices, reverse=True):
-            name = recv_names.pop(idx)
-            recv_requests.pop(idx)
-            buf = recv_buffers.pop(idx)
-
-            # FIXME: pytato shouldn't depend on pyopencl
-            import pyopencl as cl
-            context[name] = cl.array.to_device(queue, buf, allocator=allocator)
-            recv_names_completed.add(name)
-
-    # {{{ main loop
-
-    while pids_to_execute:
-        ready_pids = {pid
-                for pid in pids_to_execute
-                # FIXME: Only O(n**2) altogether. Nobody is going to notice, right?
-                if partition.parts[pid].needed_pids <= pids_executed
-                and (set(partition.parts[pid].input_name_to_recv_node)
-                    <= recv_names_completed)}
-        for pid in ready_pids:
-            part = partition.parts[pid]
-            exec_ready_part(part)
-
-            for p in part.all_input_names():
-                partition_input_names_refcount[p] -= 1
-                if partition_input_names_refcount[p] == 0:
-                    del context[p]
-
-        if not ready_pids:
-            wait_for_some_recvs()
-
-    # }}}
-
-    for send_req in send_requests:
-        send_req.Wait()
-
-    if __debug__:
-        for name, count in partition_input_names_refcount.items():
-            assert count == 0
-            assert name not in context
-
-    return context
 
 # }}}
